requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
lightgbm==4.0.0
implicit==0.7.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
openpyxl==3.1.2
pyyaml==6.0.1
scipy==1.11.0